% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callibrate.R
\name{fit_mv}
\alias{fit_mv}
\title{Fitting eir.}
\usage{
fit_mv(
  variable = "Pv_clin",
  target,
  rows,
  model = NULL,
  tolerance = 0.05,
  interval = c(1e-04, 10),
  maxiter = 10,
  verbose = FALSE,
  extendInt = "no",
  check.conv = FALSE,
  ...
)
}
\arguments{
\item{variable}{The name of the output variable to be fitted (eg: "Pv_clin" or "PvPR_LM_2_10"). This must be
a variable name in the output.}

\item{target}{The target value(s) of the specified variable. These may be a single value or vector.}

\item{rows}{The output row(s)/position(s). These may be a single value or vector.}

\item{model}{A named list of parameters to override in the
"model_parameters.txt" file}

\item{tolerance}{The desired tolerance of the fit. The fit being succesful when the \code{abs(variable - target) < tolerance)}.}

\item{interval}{The interval to be searched (see \link[stats]{uniroot} for more information, note: \code{extendInt = 'upX'}).}

\item{maxiter}{The maximum number of iterations (model runs) to search for. Importantn for keeping timescales reasonable
when fitting many sites.}

\item{verbose}{Output is printed during each iteration, including the current EIR_equil and the output target value.}

\item{extendInt}{Extend the search interval, see \link[stats]{uniroot} for more details}

\item{check.conv}{Error or warning for convergence failure, see \link[stats]{uniroot} for more details}

\item{...}{Additional arguments to \code{fit_v()}}
}
\description{
Provides basic functionality for fitting the EIR_equil parameter for a site file
to specified data (usually prevalence or incidence) at given time point(s).
}
