% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callibrate.R
\name{quiet}
\alias{quiet}
\title{Run a function silently (code from H Wickham)}
\usage{
quiet(x)
}
\arguments{
\item{x}{function call}
}
\value{
function output
}
\description{
Run a function silently (code from H Wickham)
}
