% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callibrate.R
\name{fit_v}
\alias{fit_v}
\title{Internal: Fitting EI.}
\usage{
fit_v(total_M, target, variable, rows, tolerance, verbose, model, ...)
}
\arguments{
\item{total_M}{The eir for the current model run}

\item{target}{The target value(s) of the specified variable. These may be a single value or vector.}

\item{variable}{The name of the output variable to be fitted (eg: "prev_2_10" or "clin_inc_all"). This must be
a variable name in the output.}

\item{rows}{The output row(s)/position(s). These may be a single value or vector.}

\item{tolerance}{The desired tolerance of the fit. The fit being succesful when the \code{abs(variable - target) < tolerance)}.}

\item{verbose}{Output is printed during each iteration, including the current total_M and the output target value.}

\item{model}{A named list of parameters to override in the
"model_parameters.txt" file}

\item{...}{Additional arguments to \code{fit_v()}}
}
\description{
Internal function provided to \link[stats]{uniroot} for fitting eir
}
